third_party/WebKit/LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash-expected.txt

Confirm that null is returned when an SVGViewSpec is used after its corresponding element has been removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewPreserveAspectRatio = svgView.preserveAspectRatio; is null
PASS viewTransform = svgView.transform; is null
PASS viewViewTarget = svgView.viewTarget; is null
PASS viewViewBox = svgView.viewBox; did not throw exception.
PASS svgView.preserveAspectRatioString is ""
PASS svgView.transformString is ""
PASS svgView.viewBoxString is ""
PASS successfullyParsed is true

TEST COMPLETE
